Update Documentation and LICENSE
"LICENSE" has been updated to add email address for contact
information.

The format of listing Changes in the changelog has been updated
to be more organised and readable.

"icon.png" and "icon64.png" are no longer executable.

Magali de Morangies has been added to the "package.json" as a
translator